Correctly handle JSON types in DB query in SchemaServiceImpl.findUsages() : Fixes #944
